CR: move emailOnCompletion to lib
We created the lib package for cases like this. It was empty until know but
the idea is to start moving things here as they're added to the project.
