+++ src/MCPClient/lib/clientScripts/lib/emailOnCompletion.py
